Windows 2015 warnings rollup (#1383)
* STL exporting

Only export functions, not entire class to shut up MSVC about exporting
STL containers

* windows warnings

* cmake warning

* my current configuration

* more windows warnings

* missing PDAL_DLL decl

* Make sure to initialize Everything union

* Only filters.chipper uses lround. Move it into that filter instead of globally

* hobu's current cmake configu that works

* windows warning patrol

* Refactor pg_query_once to return string instead of char* which we end up leaking

* warnings about DLL import redefinition

* unused var

* mismatched types warning

* comparison warnings

* I build project files now

* silence PyTuple_GetItem argument warning

* conversion to doubles warnings

* inconsistent dll linkage w/ python

* more accurate test output

* don't explicitly initialize Everything

* truncation warning in OCI driver

+++ src/SpatialReference.cpp
@@ -362,7 +362,7 @@ int SpatialReference::calculateZone(double lon, double lat)
     // Everywhere else.
     else
     {
-        zone = floor((lon + 180.0) / 6) + 1;
+        zone = (int) floor((lon + 180.0) / 6) + 1;
         if (lat < 0)
             zone = -zone;
     }
